third_party/blink/web_tests/external/wpt/audio-output/selectAudioOutput-permissions-policy.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] Default "speaker-selection" permissions policy ["self"] allows the top-level document.
  promise_test: Unhandled rejection with value: object "NotAllowedError: Permission denied to select audio output."
[FAIL] Default "speaker-selection" permissions policy ["self"] allows same-origin iframes.
	assert_equals: Permission denied to select audio output.. expected "#OK" but got "NotAllowedError"
[FAIL] permissions policy "speaker-selection" can be enabled in cross-origin iframes using "allow" attribute.
	assert_equals: Permission denied to select audio output.. expected "#OK" but got "NotAllowedError"
Harness: the test ran to completion.
